requirements.txt
streamlit==1.38.0
pandas==2.2.2
fpdf2==2.8.1
